backend update
1. provide different folders for different user and different content
2. get role by roleId
3. get user chapter by userId and courseId
4. change the content of getUserCourse

+++ src/main/java/cn/daringduck/communitybuilder/repository/CourseChapterRepository.java
@@ -20,6 +20,6 @@ public interface CourseChapterRepository extends JpaRepository<CourseChapter, Lo
 	
 	@Modifying
 	@Query(value = "delete from CourseChapter courseChapter where courseChapter.chapter.id = ?1")
-	int deleteChapterFromCourseChapterByChapterId(int chapterId);
+	int deleteChapterFromCourseChapterByChapterId(long chapterId); 
 
 }
